odometry

+++ TeamCode/src/main/cpp/utils/drivetrain.h
@@ -7,7 +7,7 @@ struct Drivetrain {
     C_DcMotor *back_left{};
     C_DcMotor *back_right{};
 
-    float multiplier = 1.0;
+    double multiplier = 1.0;
 
     void drive(double forward_power, double turn_power) const;
     void drive(double forward_power, double strafe_power, double turn_power) const;
